+++ app/javascript/components/map.js
@@ -1,4 +1,3 @@
-import GMaps from 'gmaps/gmaps.js';
 import { autocomplete } from '../components/autocomplete';
 
 // [...]
@@ -367,22 +366,6 @@ map.setStyle('map_style');
   }
 };
 
-GMaps.geolocate({
-  success: function(position) {
-    user_lat = position.coords.latitude;
-    user_lng = position.coords.longitude;
-    // map.setCenter(lat, lng);
-    // map.addMarker({ lat: lat, lng: lng});
-  },
-  error: function(error) {
-    alert('Geolocation failed: '+error.message);
-  },
-  not_supported: function() {
-    alert("Your browser does not support geolocation");
-  },
-  always: function() {
-    alert("Done!");
-  }
-});
+
 
 export default initMap;
