+++ app/views/activities/_filter_money.html.erb
@@ -4,6 +4,8 @@
 
   <input id="ex13" name="money" type="text" data-slider-ticks="[0, 20, 50, 100, 500]" data-slider-ticks-snap-bounds="30" data-slider-ticks-labels='["$0", "$20", "$50", "$100", "$500"]'/>
 
+  <input id="user-location" type="hidden" name="user-location"
+  />
 
   <div><%= submit_tag 'Save' %></div>
